[changed] following review comments
